Parameter Automation

+++ Tests/AudioKitTests/AKAutomationCurveTests.swift
@@ -20,7 +20,7 @@ class AKAutomationCurveTests: XCTestCase {
                                                       startTime: 2,
                                                       rampDuration: 0.1)])
 
-        let events: [(Double, AUValue)] = [ (0.5, 100), (1.5, 200) ]
+        let events: [(Float, AUValue)] = [ (0.5, 100), (1.5, 200) ]
 
         let newCurve = curve.replace(range: 0.25 ... 1.75, withPoints: events)
 
@@ -40,7 +40,6 @@ class AKAutomationCurveTests: XCTestCase {
         XCTAssertEqual(newCurve.points, expected)
     }
 
-
     func testReplaceAutomationErase() {
         let curve = AKAutomationCurve(points: [ Point(targetValue: 440,
                                                       startTime: 0,
@@ -52,7 +51,7 @@ class AKAutomationCurveTests: XCTestCase {
                                                       startTime: 2,
                                                       rampDuration: 0.1)])
 
-        let events: [(Double, AUValue)] = [ ]
+        let events: [(Float, AUValue)] = [ ]
 
         let newCurve = curve.replace(range: 0 ... 2, withPoints: events)
 
@@ -62,7 +61,7 @@ class AKAutomationCurveTests: XCTestCase {
     func testReplaceAutomationAdd() {
         let curve = AKAutomationCurve(points: [])
 
-        let events: [(Double, AUValue)] = [ (0.5, 100), (1.5, 200) ]
+        let events: [(Float, AUValue)] = [ (0.5, 100), (1.5, 200) ]
 
         let newCurve = curve.replace(range: 0 ... 2, withPoints: events)
 
@@ -131,8 +130,7 @@ class AKAutomationCurveTests: XCTestCase {
                                                      rampTaper: 0.5,
                                                      rampSkew: 0.1)])
 
-        let events = curve.evaluate(initialValue: 0,
-                                    resolution: 0.1)
+        let events = curve.evaluate(initialValue: 0, resolution: 0.1)
 
         XCTAssertEqual(events.count, 10)
 
@@ -150,8 +148,7 @@ class AKAutomationCurveTests: XCTestCase {
                                                      rampTaper: 1.0,
                                                      rampSkew: 0.000001)])
 
-        let events = curve.evaluate(initialValue: 0,
-                                    resolution: 0.5)
+        let events = curve.evaluate(initialValue: 0, resolution: 0.5)
 
         XCTAssertEqual(events[0].startTime, 0.0)
         XCTAssertEqual(events[0].targetValue, 1.0)
@@ -176,8 +173,7 @@ class AKAutomationCurveTests: XCTestCase {
                                                      startTime: 1,
                                                      rampDuration: 0.0)])
 
-        let events = curve.evaluate(initialValue: 0,
-                                    resolution: 0.5)
+        let events = curve.evaluate(initialValue: 0, resolution: 0.5)
 
         XCTAssertEqual(events[0].startTime, 0.0)
         XCTAssertEqual(events[0].rampDuration, 0.5)
